KAN-490 refactor: 정기주문 페이지 폴더구조 변경
